+++ core/server/src/data-access/controllers/user/user/upload-avatar.ts
@@ -47,12 +47,16 @@ export default function makeUploadUserAvatarController({
 
       const current_bucket = _.get(exists, "avatar.bucket", "");
       const current_key = _.get(exists, "avatar.key", "");
-      const s3_params = {
-        Bucket: current_bucket,
-        Key: current_key,
-      };
 
-      Storage.deleteS3Object(s3_params);
+      const validCredentials = current_bucket && current_key;
+      if (!validCredentials) {
+        const s3_params = {
+          Bucket: current_bucket,
+          Key: current_key,
+        };
+
+        Storage.deleteS3Object(s3_params);
+      }
 
       const user_details = Object.assign({}, exists, {
         avatar: file,
